refactor(Controller): Extract RacingGame domain to controller
