+++ src/main/java/racingcar/controller/RacingGame.java
@@ -1,5 +1,9 @@
-package racingcar.domain;
+package racingcar.controller;
 
+import racingcar.domain.Cars;
+import racingcar.domain.car.RandomEngine;
+import racingcar.domain.Round;
+import racingcar.domain.Winners;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 import racingcar.utils.StringUtils;
